python-rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
psycopg2-binary==2.9.9
chromadb==0.4.22
sentence-transformers==2.2.2
colorama==0.4.6
